Add module ColorTransfer.HistogramMatching .
